requirements.txt
Django==2.2.19
django-phonenumber-field==5.0.0
pgxnclient==1.3.1
phonenumbers==8.12.19
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
